+++ CHANGELOG.rst
@@ -7,6 +7,14 @@ foursight
 Change Log
 ----------
 
+3.3.5
+=====
+
+`PR 522: Add new params to ignore uuids or reset external expset with no pub check <https://github.com/4dn-dcic/foursight/pull/527>`_
+
+* add 'uuids_to_ignore' parameter for a list of uuids to ignore and hence not warn for this check
+* add 'reset_ignore' parameter to clear the list of uuids that are ignored
+
 3.3.4
 =====